requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
anyio==4.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.0
certifi==2024.2.2
click==8.1.7
contourpy==1.2.0
crispy-bootstrap5==2024.2
cycler==0.12.1
Django==4.2.7
django-appconf==1.0.6
django-auth-ldap==4.6.0
django-crispy-forms==2.1
django-dynamic-datatb==1.0.25
django-filter==23.5
django-htmx==1.17.2
django-mssql==1.8
django-redis==5.4.0
django-remote-auth-ldap==1.0.0
django-select2==8.1.2
django-tables2==2.7.0
django-taggit==5.0.0
django-widget-tweaks==1.5.0
django-windowsauth==1.4.0
djangorestframework==3.14.0
easysnmp==0.2.6
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.109.2
fonttools==4.47.0
frozenlist==1.4.1
h11==0.14.0
htmx==0.0.0
httpcore==1.0.3
httptools==0.6.1
httpx==0.26.0
idna==3.6
ifaddr==0.2.0
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
kiwisolver==1.4.5
ldap3==2.9.1
Markdown==3.5
markdown2==2.4.12
MarkupSafe==2.1.5
matplotlib==3.8.2
mssql-django==1.4.2
multidict==6.0.5
nicegui==1.4.15
numpy==1.26.0
openpyxl==3.1.2
orjson==3.9.14
packaging==23.2
pandas==2.1.1
pillow==10.2.0
pscript==0.7.7
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==1.10.14
pygments==2.17.2
pyodbc==5.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.0
python-ldap==3.4.4
python-multipart==0.0.9
python-socketio==5.11.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.36.3
typing-extensions==4.8.0
tzdata==2023.3
uvicorn==0.27.1
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.21.0
websockets==12.0
wfastcgi==3.0.0
wsproto==1.2.0
xlwt==1.3.0
yarl==1.9.4
zipp==3.17.0
